+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/sign/SignType.java
@@ -5,6 +5,8 @@ import java.util.function.Predicate;
 import org.bukkit.Material;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
+import com.nisovin.shopkeepers.api.internal.util.Unsafe;
+import com.nisovin.shopkeepers.compat.MC_1_19;
 import com.nisovin.shopkeepers.util.java.Validate;
 
 public enum SignType {
@@ -18,7 +20,11 @@ public enum SignType {
 	ACACIA(Material.ACACIA_SIGN, Material.ACACIA_WALL_SIGN),
 	DARK_OAK(Material.DARK_OAK_SIGN, Material.DARK_OAK_WALL_SIGN),
 	CRIMSON(Material.CRIMSON_SIGN, Material.CRIMSON_WALL_SIGN),
-	WARPED(Material.WARPED_SIGN, Material.WARPED_WALL_SIGN);
+	WARPED(Material.WARPED_SIGN, Material.WARPED_WALL_SIGN),
+	MANGROVE(
+			MC_1_19.MANGROVE_SIGN.orElseGet(Unsafe.uncheckedNull()),
+			MC_1_19.MANGROVE_WALL_SIGN.orElseGet(Unsafe.uncheckedNull())//
+	);
 
 	public static final Predicate<SignType> IS_SUPPORTED = SignType::isSupported;